7/2 17:23

+++ src/main/java/com/example/maron/controller/form/UserForm.java
@@ -16,7 +16,7 @@ public class UserForm {
     private String name;
     private int branchId;
     private int departmentId;
-    private int isStopped;
+    private Short isStopped;
     private Date createdDate;
     private Date updatedDate;
 }
